requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.8
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.6.0
numpy==1.26.3
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.1.0
protobuf==5.28.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.1
pywin32==308
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.25.0
watchdog==4.0.2
Werkzeug==3.1.3
